input: allow touchpads to be disabled
For accessibility reasons, we want users to be able to disable
touchpads. For example, users who are visually impaired may want to
disable the touchpad from a keyboard/touchpad combo device to avoid
accidental clicks or pointer movements, while still using the keyboard.

This can be tested with the following ADB command:

    $ adb shell settings put system touchpad_enabled 0

(Or the same with 1 to re-enable.)

Test: connect a touchpad, disable touchpads, check it stops working.
      Turn the toggle back on, check it starts working
Test: with touchpads disabled, connect a touchpad, check it doesn't
      work. Turn the toggle back on, check it starts working
Test: with a keyboard/touchpad combo (I tested with the Lenovo P12 Pro's
      detachable unit), check that the keyboard remains usable when the
      touchpad is disabled
Flag: com.android.hardware.input.touchpad_disable
Bug: 372857036
Change-Id: Ia2701cba221c4664673836e1a8267203022c34cd