end2end: Generate Hwc3 vsync events [5/N]
When SurfaceFlinger asks for vsync to be enabled, we need to provide
periodic event callbacks through the callback interface.

This change adds a simple generator that can handle any number of
displays, with arbitrary timing. The generator is owned by the service
implementation, and forwards generated events to the client
implementation for it to send to SurfaceFlinger.

Flag: TEST_ONLY
Bug: 372735083
Test: atest surfaceflinger_end2end_tests
Change-Id: I56d7f832609a34d4dc02a4773b8b1a4a0826650d