SF: Add sfdo command to force set pacesetter display
This CL adds a new command to sfdo to forcibly change the pacesetter
display to one specified by the ID. This is useful for testing various
connected display scenarios where refresh rates differ. The selected
display will remain the pacesetter until it is either disconnected
or another sfdo command to force-set/unset the pacesetter is invoked.

This CL also adds another sfdo command to reset the forced pacesetter.
The reset command is a no-op if no display was forced as a pacestter.
Currently, the reset sets the front internal display as the new
pacesetter if the pacesetter_selection flag is disabled.

Flag: NONE adb backdoor
Test: Invoked sfdo force-pacesetter DISP_ID && reset on a foldable
Bug: 389982668
Change-Id: I2ac891c75f4346b7e11e438b8026d5c8c5a6298f