end2end: Add an Edid Builder [3/N]
EdidBuilder generates a valid, but synthetic EDID base block (block 0)
for a requested screen resolution and refresh rate. It also allows
embedding a product name string descriptor which can be used to identify
the display later, as it is parsed out by SurfaceFlinger as part of the
gui::StaticDisplayInfo data for a display.

This will be used by the fake HWC3 implementation shortly, and will also
be usedful eventually when the test later injects displays via vkms.

Flag: TEST_ONLY
Bug: 372735083
Test: atest surfaceflinger_end2end_tests
Change-Id: I9e9eb0d13d5df0af5525511b07a2acaf5e06646a