HWComposer: Add reconnect status to DisplayIdentificationInfo
A display reconnect is a special case event in which the only available
display needs to be swapped out without being disconnected. Such cases
include (but are not limited to) ATV and headless AL in the near future.

Currently, the only way to detect a reconnect is by checking whether or
not a display is already available in the list of physical displays in
SF. This is not a good representation as the existence of a
display with a newly fabricated ID could indicate an error or a bug, and
so it is important to introduce another way to identify a display
reconnect at the time of a physical display creation.

This CL introduces a signal in DisplayIdentificationInfo to notify SF
when an incoming display is in fact reconnected instead of purely
connected.

Flag: EXEMPT refactor
Bug: 420970122
Test: HWComposerTest
Change-Id: Ic31c996e5ff1caff2f8a3ab931dffb33a26f99d4