Flag: Add flag for cursor plane behavior
This change adds a flag which is used to gate behavior changes to the
cursor surface that allows it to be presented on the cursor plane.

In subsequent CLs, when the cursor_plane_compatibility flag is set,
Sprite surfaces will use BGRA pixel format, and will allocate buffers
with the GRALLOC_USAGE_CURSOR flag. These modifications enable the
requirements of the cursor plane to be met.

Flag: com.android.graphics.libgui.flags.cursor_plane_compatibility
Bug: 404877772
Change-Id: I0eee89a99f340af642fcf1de187f6e550ce99527