Enable more input tests on host
The tests on host were disabled because we couldn't find the kl/kcm/idc
files on the host machine.

In this CL, we update the code to also include the *device* files from
the out folder into the search path.

A better way to do this would be:
A) Use the "ANDROID_HOST_OUT" environment variable + update the code so
that the kl/kcm/idc files are also installed on host in addition to the
device. Doing so would require updating the definitions for
'prebuilt_usr_keylayout' inside build/soong/etc/prebuilt_etc.go to also
install files on host (currently, it's explicitly labeled as
'device-only' there.

B) Ensure that ANDROID_ROOT is set on host correctly. Then we wouldn't
need to change the existing code inside libinput.

I couldn't find a way to avoid hardcoding 'system' inside the path, but
that seems to work locally.

I'm also not sure whether there are host-only builds where this would
fail (and whether there are ways to skip those)

Bug: 253299089
Test: TEST=libinput_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST --gtest_filter="*InputDeviceKeyMapTest*"
Flag: TEST_ONLY
Change-Id: I44467a5dd113f7318240eb8b158c4ef6f999ba99