end2end: Introduce Hwc3 observer [6/N]
The HWC3 observer intercepts Aidl calls made from SurfaceFlinger to the
a target Composer implementation and visa versa, allowing the test to
know about those events without modifications to SurfaceFlinger or the
target implementation. The calls are otherwise passed on to the intended
destination unchanged.

For now the observer intercepts calls made to the fake composer
implementation, but in the future it will be used to intercept calls
made to drm_hwcomposer.

Flag: TEST_ONLY
Bug: 372735083
Test: atest surfaceflinger_end2end_tests

Change-Id: Ife22c56ed1dca1228dceff8536a7527ad60889c9