Fix LayerCallback tests
On cuttlefish the tests were skipped because of a shader compilation
error. This masked real failues in the tests because Cuttlefish does not
support native android fences, and results in a signal time of -1.

This cl fixes the shader compilation error, and skips fence signal time
checks if the device does not support native fences.

Fixes: 328119950
Test: atest LayerCallbackTest
Flag: EXEMPT tests

Change-Id: I7b85844f240e8c3795c794426b87b201ab099c68