libbinder: trusty: remove empty buffer assertion
This assertion prevented trusty RpcServer services from co-existing with
non rpcbinder Service implementations on the same event loop. The buffer
argument to on_message is non-empty in that case because a buffer is
being used to deserialize messages for those other services on the same
event loop.

Instead of panicking, we just ignore the unused argument.

Bug: 407974066
Flag: EXEMPT bugfix
Test: trusty/vendor/google/aosp/scripts/build.py \
	qemu-generic-arm64-gicv3-test-debug \
        --test="boot-test:com.android.trusty.binderRpcTest"
Test: trusty/vendor/google/aosp/scripts/build.py \
	qemu-generic-arm64-gicv3-test-debug \
        --test="boot-test:com.android.trusty.rust.binder_rpc_test.test"
Test: trusty/vendor/google/aosp/scripts/build.py \
	qemu-generic-arm64-gicv3-test-debug \
        --test="boot-test:com.android.trusty.rust.tipc.test"
Change-Id: I4d9c35a322f2901c3fa18029344cedc7cdebf3a8