binder_parcel_fuzzer: Cause FDs to be open before tracking them
On devices with tracing and kernel binder/hwbinder, there are FDs that
are opened through libbinder while the fuzzer is running. Since we want
to track them, we need to open all of those before we get the initial
count of FDs.

These are the binder/hwbinder FDs. The socket and pmsg FD from liblog.
The FD for kernel tracing.

liblog was moved to shared_libs on device because libbinder.so will log
through liblog.so and open a separate socket/FD from what we would open
in the static liblog in the fuzzer.

Flag: EXEMPT test only
Test: binder_parcel_fuzzer # on device
Bug: none
Change-Id: Ia1d5bede10cea526f750d3bc64ec68b8ca2aaa15