SF: Use pacesetter instead of active display as the "default" display
Active display as the default display does not extend very well to the
connected display world, where the active display(s) may be off while
the external display(s) are powered-on and actually active.

This CL replaces the usage of mActiveDisplayId and
getDefaultDisplayDevice* where it makes sense. It adds a new getter
method for pacesetter to the Scheduler and SurfaceFlinger. The default
display getters are renamed as getActiveDisplayDevice* to better
reflect on their source and to discourage its use as the "default".

Flag: com.android.graphics.surfaceflinger.flags.pacesetter_selection
Bug: 255635821
Test: manual testing on clamshell and foldable w/ ext display
Change-Id: I87aeb3551cf893f8161534cdd5568bb90e4f47cf