Add SimpleTransactionState and TransactionListenerCallbacks.
This is a step towards full TransactionState, which was again reverted.

SimpleTransactionState holds data that is passed by copy / move.
TransactionListenerCallbacks deals with flattened listener callbacks.
Unlike other data, its members are not also members of Transaction class
but rather are computed from the full listener callback map prior to
calling setTransactionState and so it is for the moment separated.

The reason for this granularity is to better isolate regressions should
they arise.

Flag: EXEMPT refactor
Bug: 385156191
Test: presubmit
Change-Id: Ifc4b40150dbfbbaf5bf9e28d1afd0063847675d0