SF: Break down onActiveDisplaySizeChanged()
Remove SF::onActiveDisplaySizeChanged(), and replace it with
independent calls to Scheduler::onPacesetterDisplayAreaChanged() and
RenderEngine::onActiveDisplayAreaChanged().

The main difference in this CL is that the Scheduler is fed the display
area of the pacesetter instead of the "active" display, and that the
render engine is given the area of the largest configured display for
the cache instead of the active display.

Flag: com.android.graphics.surfaceflinger.flags.pacesetter_selection
Bug: 255635821
Test: dumpsys SurfaceFlinger | grep displayArea w/ foldable
Change-Id: I724066f39289e21e395d637720bc8994f070e1bf