SF: Invoke DisplayDevice::animate*Overlay() on all displays
Right now HDR:SDR ratio and refresh rate overlays are only animated on
active displays. This CL splits animateOverlay() into two, one for each
of the overlay, and invokes it for for all physical and virtual
displays in SurfaceFlinger. The animate function call should check for
presence of the overlay so no pre-check is required.

Flag: EXEMPT trivial refactor
Bug: 255635821
Test: Tested on a foldable with hdr/sdr ratio using made you look
Change-Id: I68f76bcbaecd2f0a1b429160fb6c119a6738ed8f