Crash if can't add window infos listener
When the listener can't be added, device will be unusable. Today, the
callers of 'addWindowInfosListener' simply ignore the condition where
the listener can't be added.

In practice, there's nothing that the client can do to recover from
this. Reboot is the best remedy for those cases.

In the future, this check will help simplify the handling of the initial
condition of window infos.

Bug: 404661556
Flag: EXEMPT refactor
Test: none
Change-Id: I5f887c4a50f2b201ed90025aac468e41010cd12c