Force callers to consume initial WindowInfosUpdate
Previously, this argument was optional. That suggested that there's
guarantee that the update will be provided soon after the listener is
registered. However, that's not the case.

In this CL, we are forcing all users of WindowInfosListener to consume
the initial data. This will save engineering time by forcing the
developer to consider how to handle initial condition immediately, and
not have to do run-time debugging to figure out why the window infos are
sometimes empty (and sometimes not).

Also in this CL, remove the std::pair construct in favor of the existing
WindowInfosUpdate.

Bug: 404661556
Flag: EXEMPT refactor
Test: none
Change-Id: I7ead3f2846fe722f7e7af656feb1a0782fe364ab