Return WindowInfosUpdate when registering infos listener
In this CL, we refactor WindowInfosListener to return a Result-wrapped
WindowInfosUpdate. This makes it more user-friendly.

As an alternative, we could further refactor this by never returning an
invalid WindowInfosUpdate - since failure to listen to windows would
mean that the device is not usable. To ensure that the initial info is
consumed, we could add a "no-discard" annotation to that function.

Bug: 404661556
Flag: EXEMPT refactor
Test: none
Change-Id: I16d13919e92aa4b3ee0ac5f6024c7552ebb0da49