libgui: Make Surface_test more robust and improve logging
In b/365708032, we're seeing flakiness around the creation of a
SurfaceComposerClient, even in tests that don't use it. The class is not
being initialized since SF is likely not available

This is only used by a small number of the 100+ tests in this file, so
if the tests don't rely on it, we can avoid the initialization.

Bug: 365708032
Flag: EXEMPT adding logging
Test: atest libgui_test
Change-Id: I60c7d9efd5e61fb341298cff3cb8847eb2a6b8ba