EventHub: Save configurationFile to Device
Previously, in <36b1159>, the configurationFile was inadvertently
ommitted as a member of a newly created Device.  This commit has
`loadConfiguration()` add the configurationFile used for the Device to
the PropertyMap with the key "configurationFile".  In the dumpsys
output, this value will be retrieved from the PropertyMap.
Additionally, the configurationFile member has been removed from
the Device struct.

Bug: 407089480
Flag: EXEMPT bugfix
Test: Build and flash brya, execute `dumpsys input`, verify that the
`ConfigurationFile:` field is no longer empty for a device that uses an
device configuration file. For those devices that do not have a device
configuration file, it says "<none>" to match the format in the other
fields.

Change-Id: I25cdaa91b8656838c522b896bd34396a70496c36