Support multiple sensors in AOSP fusion algorithms
In situations where multiple sensors of the same type are active, only
one must used for running sensor fusion algorithms otherwise they will
produce invalid output. Ensure that the first sensor of each type is
selected, and also ensure that if other sensors of the same type are
active, their sensor samples are not used in the fusion algorithms.

Fixes: 408064104
Test: manual validation
Flag: EXEMPT bugfix
Change-Id: I1497224b18a2e6cef6b4a4b863a8615504d9a356