SF: Subsume frametimeline namespace in scheduler
TokenManager logically belongs as part of the scheduler. Once it is
removed from FrameTimeline it becomes apparent that there really is no
need for a frametimeline namespace and having it around only makes the
code integrating the timeline and token components more difficult to
read and write due to navagating around the namespaces.

Bug: 414584347
Test: none
Flag: EXEMPT, refactor
Change-Id: I9bf32fceffc019c0fa10a0082505bc9a365068c7