end2end: Remove DisplayIdentification dependency
EdidBuilder_test currently depends on ui/DisplayIdentification.h, but it
has to be moved back to SF. Once it is moved, the end2end test framework
will not be able to depend on it, at least not in a trivial way.

Remove all DisplayIdentification dependency from end2end
EdidBuilder_test by deleting the tests that use EDID parsing logic.

Bug: 415035569
Flag: EXEMPT refactor
Test: surfaceflinger_end2end_tests
Change-Id: I82f46c0a55beab04bc80d7af6740e67aa11389ad