Add libbpf_prog gpuWork.bpf target
kMaxTrackedGpuIdUids is used for the number of entry of bpf map which
caused "variable length array folded to constant array" error.
This CL fixes the issue by using #define instead of static const.

This is a preparation for migrating to the Rust platform bpfloader.

Flag: EXEMPT gpuWork.bpf is not used
Bug: 410982483
Test: m gpuWork.bpf
Change-Id: Icdf08b8eee39c98a0ab88e887109632d62de10e1