Add FenceTime::wasPendingAt()
FenceTime::wasPendingAt() takes a time as an argument, and returns a
bool indicating if the fence would have been pending at that time.

Time specified can only be in the past. If the fence is currently
pending, it is assumed that fence is pending at any given time.

Flag: EXEMPT PURE_REFACTOR
Bug: 409682021
Test: New FenceTime_test
Change-Id: I206a5979f0b26a2c11f633b5d1894a0ee2db448e