Use current thread for SetTaskProfiles
After the recent std::thread refactor, the input thread profile is being
set to the current thread. Therefore, we no longer have to explicitly
provide a thread id to the SetTaskProfiles - we can just send "0", and
rely on the fact that ExecuteForTask calls GetThreadId() for those
cases. That's not documented, but is an existing behaviour.

This will be supported on more platforms, and allows us to remove the
__ANDROID__ ifdefs.

Bug: 378761104
Test: none
Flag: EXEMPT refactor
Change-Id: I8e8325e1ecb43d31621657d250530924f40ecc43