EventHub: Log slow reads
Logs slow read times.

A 'slow read' is when the timestamp reported by the kernel is lagging
more than 100ms behind the time when it is being read by
inputflinger.

This will give an indication of if/when input processing in the
InputReader thread is backed up.

Test: Presbumit
Flag: EXEMPT logging only
Bug: 414757227
Change-Id: I73132f09cba414b8145d74c635efbd3088e3b225