VirtualDisplays: Create an updated Virtual Display implementation.
This implementation simplifies the buffer tracking logic, increasing
maintainability. It should function as a drop-in replacement for the
other.

This CL implements a single threaded version that wastes a lot of
buffers due to never dequeuing any from the sink, which is not
shippable.  But it's a start. A subsequent CL splits out an extra thread
for interactions with the app process and better buffer management.

This CL also constantly repeats the same HWC slot id and sends a buffer
every frame to HWC for HWC-backed virtual displays. This is also
addressed in a subsequent CL.

See go/wb-virtual-displays

Bug: 340933138
Flag: com.android.graphics.surfaceflinger.flags.wb_virtualdisplay2
Test: atest, presubmit, new tests

Change-Id: I04b2b1dec6a55255aeeb54015204d58b3bf97174