Convert motion flags to use ftl::Flags
This will help ensure correctness of the flags, and will allow us to
eventually move away from the raw flags.

It will also make it more clear that the Java-based intdefs (to be
defined in a separate CL) are related to these native flags.

Things like NotifyMotionArgs and all key-related objects are left
unconverted, as an exercise for the reader.

Bug: 245989146
Test: none
Flag: EXEMPT refactor
Change-Id: I1af7f43f19c22cefdd710b18bfa3938024832ae0