Fix set initialization for snapshot requests
The set for exclude layer IDs was always initialized with 1 element
containing the value "0" because of the double curly braces used in the
snapshot request initialization. This led to region sampling always
doing additional unnecessary work when getting layer snapshots since the
content of the set is used to determine which "getSnapshot" path is
taken.

Bug: b/417069600
Test: logging, traces
Flag: EXEMPT, bug fix
Change-Id: I79e1b012888c4210e525896e465f3b9780a9937c