Surface: Improve logging dramatically
Adds a bunch of SURF_ALOG* macros that will automatically log the name
and ID of the underlying IGBP. Also grabs those values at connect time.

This should make it clear when multiple surfaces are being interacted
with, which ones are affected by various operations.

When a surface is disconnected, the name is suffixed with -DISCONNECTED
so that subsequent operations on a dead surface can be understood
better.

Bug: 413059222
Bug: 418318000
Bug: 420318909
Bug: 421921018
Test: n/a
Flag: EXEMPT adding logging

Change-Id: I510d964e7aae0ec4a723f50a0308ce151566c2e5