libbinder-trusty: Use trusty_peer_id directly in Rust
Change the C bindings exposed to Rust to pass a trusty_peer_id by value
instead of passing a byte slice. This lets us avoid some unsafe
(unpacking the bytes) and makes it so the Rust and C++ RpcServer
implementations are using ABI compatible peer IDs, so a Rust server
would be able to use a per-session callback defined in C++ and vice versa.

Bug: 401776482
Test: ./build-root/build-qemu-generic-arm64-test-debug/run --headless --boot-test "com.android.trusty.rust.authmgr_be_lib.test"
Flag: EXEMPT bugfix
Change-Id: I14a8596fed7da46fd1d9c62c9b9437e82530a31d