RPC Binder: ABinderRpc_Accessor is not RefBase
This object inherited from RefBase, but it is only/
constructed/deleted. This creates a leak and logspam,
and we do want all RefBase objects to be managed with
RefBase, otherwise if they are accidentally passed
into sp<> constructors (in libraries that use the
default implicit constructor), it could accidentally
create double-ownership. So removing RefBase parent
class since it is not needed.

Fixes: 419661001
Test: binderRpcTest (with RefBase warnings as fatal errors)
Change-Id: If6cb6f401986e9f09509e00f61b22146b23b4408