Binder: Update zeroMemory for AIDL transactions
Due to compiler optimizations, there are cases where memset is
optimized out of the actual compiled artifact if the compiler
knows that memory cell won't be access by the process. For
security reasons this isn't ideal, this change introduces
memset_explicit and an assembly instruction to match memset_explicit
as a fallback.

Bug: 416248344
Test: aidl_unittests, aidl_integration_test, binderLibTest
Flag: EXEMPT bugfix
Change-Id: I596229d97338cfc280fb87dc02f75705ea33b273