DualDisplayTransactionTest: init ID vars via getters
When adding a FlagManager check in DisplayIdentification, and then a
test in HWCompserTest.cpp, a linker order issue arose where
DualDisplayTransactionTest's static variable initialization is dependent
on FlagManager, which did not have a chance to initialize its own
globals yet.

In addition, it turns out that  the aconfig flagging library itself
generates a global (provider_) with undefined state.

In the meantime, to solve this, we initialize kInnerDisplayId and
kOuterDisplayId constants statically via getter methods instead of
during initialization of DualDisplayTransactionTest's object.

Flag: EXEMPT refactor
Bug: 421192173
Test: libsurfaceflinger_test
Change-Id: I4aef05def1ed7d9f85d1606076994a23993601e8