RPC Binder: -ECONNRESET => DEAD_OBJECT
This is for consistency with kernel binder. Sometimes, socket
failures result in -ECONNRESET when the connection is not
available. Right now, this may be thrown as a different error,
and we will only recover after the next transaction returns
the DEAD_OBJECT error. Coalesce this error so that we have
consistent behavior between kernel and RPC binder.

Bug: 392717039
Bug: 419364025
Test: binderRpcTest
Flag: EXEMPT .
Change-Id: I80cf8124f9d9c4749b556a9287720f985cf82c3e