SF: Avoid redundant pacesetter setting in SF::setPhysicalDisplayPowerMode
SurfaceFlinger::setPhysicalDisplayPowerMode() currently sets the front
internal display as pacesetter and invokes
SF::applyRefreshRateSelectorPolicy() every time. This is wasteful if the
pacesetter designation won't change.

This CL adds a check to the call to applyRefreshRateSelectorPolicy() and
Scheduler::setPacesetterDisplay() so that they are only invoked when
there's actual change to be made.

Flag: com.android.graphics.surfaceflinger.flags.pacesetter_selection
Bug: 389983418
Test: SetPhysicalDisplayPowerModeTest
Change-Id: Iac4b00fec08c5ae9b7aa97200d04194edd793348